first refactor commit, add views folder and start moving across page routes for landing
