+++ socioprophet-web/client/src/views/legal/Privacy.tsx
