+++ socioprophet-web/client/src/views/legal/Terms.tsx
