+++ socioprophet-web/client/src/views/legal/scss/terms.scss
