+++ socioprophet-web/client/src/views/not-found/NotFound.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Header from '../header/Header';
+import Header from '../../components/header/Header';
 
 import './scss/notFound.scss';
